pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "leocec"
version = "0.1.0"
dependencies = [
  "dbus-next",
  "pyserial"
]

[project.scripts]
leocec = "leocec.__main__:run"
